+++ click_extended/decorators/convert/__init__.py
@@ -1,7 +1,11 @@
 """Initialization file for the `click_extended.decorators.convert` module."""
 
+from click_extended.decorators.convert.convert_temperature import (
+    convert_temperature,
+)
 from click_extended.decorators.convert.convert_time import convert_time
 
 __all__ = [
     "convert_time",
+    "convert_temperature",
 ]